Tweak organization and documentation.
