+++ cli/src/cmds/find_repo.rs
@@ -1,7 +1,7 @@
 use std::{env, path::Path};
 
 use rsgit_core::repo::Result;
-use rsgit_on_disk::OnDisk;
+use rsgit_on_disk::OnDiskRepo;
 
 // Discover a git repo starting from the given path.
 //
@@ -14,13 +14,13 @@ use rsgit_on_disk::OnDisk;
 // where there is a `.git` directory nested within the
 // given path.
 //
-// Returns a `Result` with `rsgit::repo::OnDisk` or
+// Returns a `Result` with `rsgit::repo::OnDiskRepo` or
 // `rsgit::repo::Error` if no such repo exists.
 #[allow(dead_code)] // TEMPORARY: Until other code actually uses this.
-pub fn from_path<P: AsRef<Path>>(path: P) -> Result<OnDisk> {
+pub fn from_path<P: AsRef<Path>>(path: P) -> Result<OnDiskRepo> {
     // TO DO: Look in other places for repo.
     // https://github.com/rust-git/rsgit/issues/80
-    OnDisk::new(path)
+    OnDiskRepo::new(path)
 }
 
 // Discover a git repo starting from the current working directory.
@@ -34,11 +34,11 @@ pub fn from_path<P: AsRef<Path>>(path: P) -> Result<OnDisk> {
 // where there is a `.git` directory nested within the
 // given path.
 //
-// Returns a `Result` with `rsgit::repo::OnDisk` or
+// Returns a `Result` with `rsgit::repo::OnDiskRepo` or
 // `rsgit::repo::Error` if no such repo exists.
 #[allow(dead_code)] // TEMPORARY: Until other code actually uses this.
 #[cfg(not(tarpaulin_include))]
-pub fn from_current_dir() -> Result<OnDisk> {
+pub fn from_current_dir() -> Result<OnDiskRepo> {
     // This function is excluded from code coverage because we can't
     // be sure of the execution environment while testing. So we keep
     // it as simple as possible.
